logger/requirements.txt
flask==1.1.2
flask_migrate==2.5.3
flask_script==2.0.6
pika==1.1.0
psycopg2-binary==2.8.4
pyyaml==5.3.1